+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 ## v10.3.0 (2021-04-22)
 
-* Updates Instabug Android and iOS SDKs.
-* Migrates iOS to XCFramework.
+* Migrates iOS to XCFramework
+* Adds new API: Replies.showNotificationAndroid
 * Deprecates Instabug.setVideoRecordingFloatingButtonPosition in favor of BugReporting.setVideoRecordingFloatingButtonPosition.
+* Updates Instabug Android and iOS SDKs
 
 ## v10.0.0 (2021-02-16)